Copied commit from MapML-Leaflet-Client:
In mapml.js:

Rename M.MapMLLayerControl._onMapMoveEnd to _validateExtents so as to be 
less generic and more descriptive of what goes on when the moveend event 
is handled.

Instead of generating moveend events from within the layer control, 
directly execute or handle the more appropriate and 'extentload' 
custom event handler, newly renamed _validateExtents method (see above).

In map-layer.html: 

Rename _onMapMoveEnd to _validateDisabled, so as to be more specific
about what the method does, which is to check the 'disabled' property
against whether the layer is a) intersecting the map frame and b) of the
same projection as the map and c) has a zoom 'extent' range in which
spans the current zoom level of the map.

Stop generating map moveend events. Instead, execute the handler 
(the newly renamed _validateDisabled) directly.

Move the location of the addition of the _onLayerChange handler to the
layer until after the layer is added to the map.  The reason was that
the handler was immediately invoked by the addition of the layer to 
the map which caused the internal property _apiToggleChecked's default
value to be reversed.  That property is kinda complex and is needed
in order that the Polymer observer functions behave correctly.

In web-map.html:

Stop generating moveend events. Set up the layers' _validateDisabled
internal member as the handler for moveend events.