Created tests for styleSheet loading, removed deprecated action attribute code.  (#101)
* Merge origin/master

Conflicts:
	package.json
	test/karma.conf.js
	test/layerTests.js

* Merge origin/master

Conflicts:
	package.json
	test/karma.conf.js
	test/layerTests.js

* playwright testing added, karma testing moved to tempTests

* using Jest testing framework

* added express server for development testing

* jest fix

* server files neatness

* server.js moved

* tests run one by one

* readded karma for client side unit tests

* jest framework fix, karma removed

* clean up

* converted rest of the test

* Fix drag and drop issue #96

* Fix karma issue

* Fix karma firefox launcher issue

* needed for merge with upstream

* update with upstream and remove karma config

* reformatting for pull request

* reformatting

* reformatting part 2

* Fix some of the  messages from eslint processing, using es v 6 setup in 
gruntfile.js

* Remove duplicative code that supports the extent action attribute. This
attribute was an early attempt in MapML to encode map state semantics into 
markup, which was later replaced with the more effective URL template 
mechanism applied in Leaflet and other libraries.

* Fix bug in which zoom input was being copied twice into extent.template.values
property.  Don't know if there was any expression of the bug, Ahmad Ayubi
noticed it and suggested the fix.

* WIP - creating/copying <link rel=stylesheet> or <style> from mapml content
that is remote or inline.  See comments in addData function for further
ideas.  Adds a couple of test files that can be used as a basis for future
tests.

* Continue to remove code that supports the extent@action attribute.  Retrieve
some code that was  associated originally with the code path for @action-based
extents, but had not been incorporated into the more recent template-based
code path.  The removed code that is being relocated supports static <tile>
elements in the <layer->content</layer->.  Add local test html file with this
configuration, which could be improved for tests yet to be written:
static-tile-elements.html

* Add some potential test html cases to test the copying of inline and 
linked stylesheets from inline <layer->content</layer-> to an appropriate
location within the shadow tree, that should hopefully get garbage-collected
with the appropriate life cycle of the container's content into which they
are copied. See comments in these documents

* Potentially fix the container location of link and style copied from
mapml documents loaded from the network (currently > 1 page is loaded, if
there are 'next' links in the document). The link and style elements are
recreated in the container for the layer, but they are not 'garbage collected'
upon map moveend events, despite a one-time handler function being registered.
Why is that ???

* Continue to remove code that supports the extent@action attribute.

* Added tests for mismatched layers and the order style tags and links are added/added at all

* Add tile caches for testing

* Add vector tile caches for testing. Fix jshint lint error. Fix fetch handler
which was requiring text/mapml as a content type return.  Seems a bit strict
at this point, especially because can't readily control the content types in
the built-in ide http server, perhaps also true of test http server.

* Move data directory into e2e

* Move data directory into e2e

* style order fix

* moved data folder for fetch from peter's repo

* Add parsing of <link rel=stylesheet> and <style> to M.MapMLLayer._initExtent,
so that the layer will have a layer._stylesheets property if the parsed
MapML document contains <link> or <style> elements.

* integrated peters cached tile data and test files

* added tests for vector tiles

* Fix bug revealed by excellent test of inline and remote stylesheets in
mapml vector tiles.

* test name changes, redefined expected outcome

* Fix for loop indexing cut and paste error

* Fix issue with base resolution in tile rendering of linked stylesheet

* styleorder test checks for order now

* Removed unused event handlers for mapml tiles in MapMLFeatureLayer addData

* now runs tests on firefox

* Create M.Util.parseStylesheetToHTML(mapml, base, container), used in 3 places.

* Remove CSS on moveend event for MapMLFeatures

* Create M.Util.parseStylesheetToHTML(mapml, base, container), used in 3 places.

* Create M.Util test suite

* Remove extra commas in jest.config.js

* Added additional Util Tests

* Added edge case handling to parseStylesheetAsHTML

* Fixed a test title

* Added jshint suggestions

* Logic change when handling null arguments for parseStyleSheet

* Tests for M.coordsToArray Function

* Updated jest config

* Removed wekit from jest config and changed travis dist to bionic

* Travis dist change

* Added GUI function for travis ci

* Jest Browsers set to just chromium

Co-authored-by: Peter Rushforth <peter.rushforth@canada.ca>
Co-authored-by: Peter Rushforth <peter.rushforth@gmail.com>