Fix opacity when the attribute is provided at page load, _opacity is not set (setter not run), should return the attribute value if _opacity is nullish.
Update to do list

Work in progress - BNG and Feature Links (BC link) seem to have
mutually exclusive solutions - zoomTo after changing projection on
map is necessary to inform Leaflet of the changed CRS (scale difference)
but the history gets messed up and the BC link doesn't quite render
(you have to generate a moveend on the map, then it'll draw).

Removed a section  of code from _handleLink that sets the map viewer
projection when there is a single map layer and it has a different projection
from that of the map BECAUSE this is handled by the MapMLLayer in
selectAlternateProjection and we don't want to trigger the promise
involved in disconnecting / reconnecting all map layers (happens when
mapml-viewer projection attribute changed callback happens).