Refactor layer.js._calculateExtent into layer.js.extent + MapMLLayer._calculateBounds + map-extent.js._calculateBounds
Fix problem with MapMLLayer._calculateExtent that wasn't calcing the
zoom bounds properly.  Problem visible on main when clicking on Finland
or Sweden in Arctic SDI.

Add some guard conditions to AnnounceMovement.totalBounds, which was
firing too early in the map's life cycle and was causing issues with
layers and descendant elements being not ready.

Make layer.whenReady() a bit stronger, by determining if the shadowroot
is ready too, which may head off some timing problems.