WIP on map-link
- add extentsReady as a criteria for layer initialization to proceed.
the problem encountered is that the `checked` attr. change callback
does a layer-.whenReady() which happens much earlier and doesn't take
into account the extents' readiness, so some properties of _extentLayer
aren't set up when the layer-._layer is added to the map, resulting in
console errors.
- the MapMLLayer.copyContentToShadowRoot method was setting the href
attribute of map-link elements erroneously.  Remove that code,
replaced the map-link.href getter with a better approach.
- remove code from MapMLLayer that sets up the zoomin/zoomout links,
move it to map-link href getter, essentially.
- change MapMLLayer._validateLayerZoom() to use map-link, and 
not _properties.zoomin / zoomout
- comment out map-link._createZoominOrZoomoutLink() since it isn't 
necessary (perhaps), being handled by the href getter.