WIP on map-link really really WIP
- add layer.#hasConnected private so that can avoid attributeChangedCallback
calls queued by browser on upgrade etc.  Allows to remove individual
this.whenReady() calls and consequent microtask queue enlargement
- remove use of inducing layer.connectedCallback by the 
layer.src.attributeChangedCallback - only do the side effects which
means calling the functions that layer.src.connectedCallback calls,
seems to prevent inadvertent duplication of content in layer.shadowRoot
- per advice from HTML Standard, create layer.shadowRoot in constructor
as well as initialization of opacity.
- unconditionally purge layer.shadowRoot during onRemove
- move selectAlternateOrChangeProjection() to after copying content
to shadow root, insert into next then in promise chain which waits
on all that contents' whenReady() to settle which can only be done
there because it needs access to the map-link.getBase() instance
function to determine the correct URL to load, which isn't a function
until the map-link and all its "sibling" content becomes ready.
- add changeprojection messaging to fetch catch block, but don't call
it a "_fetchError" because that's confusing (not really a fetch error).
Also, the overall promise should fulfill, just not until after changing the src
or mapml-viewer.projection as selected.
- get rid of copy of getBase() on layer, which was cut-and-pasted from
map-link, and can't work on layer because the document context of
the map-link is element instance-based and not available to layer.
- remove the parameter from layer.getProjection(), make it detect whether
the content is remote (.src exists, so .shadowRoot should contain stuff)
- change other stuff that relies on layer.shadowRoot to use layer.src
to detect if layer *should* have a shadowRoot
- make map-link.whenReady() reject if rel=alternate but no href nor
projection
- attempt to map-link.whenReady() to reject with message when await
used on parentExtent.whenReady() and _templateVars.inputsReady