WIP on map-link:
- adapt FeatureLayer so that it doesn't expect feature data
during initialization.  Change options.static to be options.tiles and
let it be false by default: FeatureLayer is used only for its addData
function by TemplatedTileLayer, in which case and only in that case
options.tiles is passed as true; 

- adapt TemplatedTileLayer only uses M.featureLayer for access to 
the addData function, which it now calls once for each feature found 
in the fetched MapML vector tile response; map-feature elements are not
added to the DOM for each feature, only their geometry is rendered as
SVG <g> and <path> elements.

- adapt TemplatedFeaturesLayer - remove use of options.static which now
has a default of false

- remove use of M.featureLayer.options.static by MapMLLayer._mapmlvectors,
because it now has a default of false (and is more correctly called
options.tiles).

- remove QueryHandler.featureLayer.options.static, because that option
is false by default and has been renamed to options.tiles

- update map-feature to extend the bounds of the M.featureLayer to which
it is added, since the bounds of the Leaflet layers are harvested to
calculate the extent / bounds of the parent <layer-> element.