WIP on map-link:
- add console log of document that causes parser error, useful for 
debugging in geoserver query layers, which serves errors as xml documents

- change layer-.zoomTo() use of whenElemsReady() to use this.whenReady()
because the layer is only ready after the elements are ready, so 
whenElemsReady() doesn't wait long enough in some cases.

- add M.geometry.layerBounds and M.geometry.zoomBounds via passed options
of the same name.  Necessary so that the M.featureLayer._layers hash
table has a value of these properties for every layer it holds, which
is necessary when setting the bounds of the overall <layer-> 

- make map-feature getters for zoom,min, max return fallback values if
the attribute isn't set

- map-feature disconnectedCallback can happen when ancestor elements
are disconnected, which may have cleaned up  resources used by map-feature,
so the disconnectedCallback should test if they've been dealt with
already and not throw errors if so.

- rename-refactor M.FeatureLayer.addData to M.FeatureLayer.renderToLeaflet

- fix bug in map-feature._updateExtent (used non-uppercased comparison
for tagName value, which returns uppercase always)

- add map-feature.extent.zoom property to match that of layer-.extent

- add map-feature.getMeta API which returns the appropriate map-meta
element in the ancestors of the map-feature, if any

- update M.FeatureLayer:

- add addLayer(layerToAdd) which conforms to L.LayerGroup API
- update removeLayer(layerToRemove) to maintain internal structures of
M.FeatureLayer, including .layerBounds, .zoomBounds and ._layers
- make removeLayer return this, conforming to L.LayerGroup API

- refactor geometryToLayer to tag it as 'private': _geometryToLayer

- when feature is queried or static, we need to extend the M.featureLayer
layerBounds and zoomBounds

- for QueryHandler:

- synthesize a <map-geometry><map-point> for all features returned by
query if they do not posses a <map-geomtry>. Uses :has() because this
was recently deemed fit for use across browsers, with FF 121

- change approach to parsing: text/mapml documents are handled the same
way, but mapml documents returned under a different media type should
get handled ok.  text/html documents will get parsed and appended to
a synthesized map-feature>map-properties element, which should display
them as parsed HTML.  Other text documents or unrecognized stuff will
display as text in the popup, hopefully.