WIP on map-link axisInferring.test.js
- update map-feature.js - when calculating extent, don't use the zoom
 fallback value unless no zoom attribute exists. map-feature.zoom
 looks for map-meta in the parentage of the element, but falls back 
to the projection, which is not suitable for calculating extents when
the coordinates require a zoom value for calculation

- rename M.getBounds to M.getBoundsFromMeta

- add map-meta[name=extent]-based bounds as default layer bounds for
MapMLLayer._calculateBounds, so that map-meta, _mapmlvectors, 
static tiles and map-extent -generated bounds are additive

- update Util.pointToPCRSPoint and Util.boundsToPCRSBounds logic so
that zoom parameter is optional for points that don't require a zoom
to transform i.e. gcrs to pcrs, pcrs to pcrs don't need a zoom value

- update Util.getBoundsFromMeta to throw where it is invoked without
a zoom where a zoom value is required to convert coordinate system
i.e. tile,tcrs,tilematrix

- remove fallback cs from Util.getBoundsFromMeta since it could be 
wrong

- remove fallback to zoom = 0 where zoom parameter is not provided,
throw instead where it could cause a problem

- update axisInferring.html and .test.js.  Add (commented) features
that represent the bounding box of the map-meta[name=extent], for
debugging purposes

- update the expectations to expect that the extents of the layers
involved should include the map-meta[name=extent] bounds AND the 
contents of the layers.