Update map-feature.getFallbackCS()
Remove 'meta' option for showPaginationFeature.
Add meta array property to each feature that is
returned by a query.  The meta array points
to an array of <map-meta> elements that were
retrieved from the query response document.
These are appended to the <map-link> shadow root
for each pagination feature, which might be a
bit inefficient but will not give incorrect results.

Make it so that <map-extent>.getMeta looks for
name=cs type metas.

Add queryFeatureCSFallback test